server/requirements.txt
Flask==1.1.2
psycopg2==2.8.5
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.16
python-dotenv==0.13.0
